modified script to start up services

+++ access-point/scripts/start-ap-services.lua
@@ -1,4 +1,4 @@
-package.path = package.path..';/portal/lib/?.lua'
+package.path = package.path .. ';/scripts/lib/?.lua'
 local nixio  = require 'nixio'
 local reload = require 'reloader'
 local fs     = require 'nixio.fs'
@@ -10,20 +10,24 @@ if s ~= 0 then
 end
 
 -- kill hostapd support if needed
-local check = '/usr/bin/test -e /tmp/hostapd.support.pid'
-local s = os.execute(check)
-if s == 0 then
-  local cmd = '/bin/cat /tmp/hostapd.support.pid'
-  local pid = io.popen(cmd):read('*l')
-  local cmd = '/bin/kill %s'
-  local cmd = string.format(cmd,pid)
-  local s = os.execute(cmd)
-  if s ~= 0 then
-    nixio.syslog('info','failed to kill hostapd support. Exit code: ' .. s)
-  end
-  local s = fs.remove('/tmp/hostapd.support.pid')
-  if s ~= true then
-    nixio.syslog('err','failed to remove /tmp/hostapd.support.conf')
+local cmd = '/usr/bin/test -e /tmp/nointernet'
+local internet = os.execute(cmd)
+if internet ~= 0 then
+  local check = '/usr/bin/test -e /tmp/hostapd.support.pid'
+  local s = os.execute(check)
+  if s == 0 then
+    local cmd = '/bin/cat /tmp/hostapd.support.pid'
+    local pid = io.popen(cmd):read('*l')
+    local cmd = '/bin/kill %s'
+    local cmd = string.format(cmd,pid)
+    local s = os.execute(cmd)
+    if s ~= 0 then
+      nixio.syslog('info','failed to kill hostapd support. Exit code: ' .. s)
+    end
+    local s = fs.remove('/tmp/hostapd.support.pid')
+    if s ~= true then
+      nixio.syslog('err','failed to remove /tmp/hostapd.support.conf')
+    end
   end
 end
